requirements.txt
FastAPI>=0.104.0
uvicorn>=0.24.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
python-jose>=3.3.0
passlib>=1.7.0
pytest>=7.0.0
httpx>=0.25.0
aiofiles>=23.0.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
